Updated the buy methode

+++ vending_machine_library/Models/Transaction.cs
@@ -6,11 +6,11 @@ namespace vending_machine_library.Models
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
 
         public Transaction() { } //def constructor
 
-        public Transaction(int id, string productName, float price)
+        public Transaction(int id, string productName, double price)
         {
             Id = id;
             ProductName = productName;
